C# & .NET Framework Course Exercise Pg 114
Created an application that takes two integer inputs from the user - the second input being optional. If the user does not provide a second integer input, the program has a default value for the second parameter. 

Currently non-functional: Issue with the inability to move to default 2nd value if nothing is entered/Enter key is pressed during the Console.Readline() event.