C# & .NET Framework Exercise Pg 114
Created a code that takes two integers as parameters, where the second int parameter is optional for entry. 

Currently functional with exceptions -  trying to figure out how to make it work if no user input is given for the 2nd parameter request is given, attempting several different methods to no avail at the moment.