C# & .NET Framework Course Exercise 112
(rework) Created a console app that took two Int parameters, ran them through a void method that conducts a mathematical operation on them, but only displays the 2nd int parameter, regardless of the result of the math operation. 

Reworked the code to function by commenting-out the "void" and adding a datatype for the method, as well as adding a return for correct output.