Update Program.cs
(rework) Redesigned code with better understanding of the if statement/try.Parse combination. Previous version forced defaultNum02 to use 10 regardless of if a user input anything or not. In this case, a user can enter a valid Integer, and if not, if they put in something like 'asdf', the default value for defaultNum02 will be 10.